Extract DealDetail; update home page; trim Privy config
