+++ src/lib/privy/config.ts
@@ -3,16 +3,10 @@ import { base, baseSepolia } from "viem/chains";
 
 export const PAYMENT_CHAIN = baseSepolia;
 export const PAYMENT_CHAIN_NAME = PAYMENT_CHAIN.name;
-
-/** Payment chain id (number). Use for switchChain and comparisons. */
 export const BASE_CHAIN_ID = PAYMENT_CHAIN.id;
-
-/** CAIP-2 chain id for the payment chain. */
 export const BASE_CHAIN_ID_CAIP2 = `eip155:${BASE_CHAIN_ID}` as const;
 
-/**
- * Returns true if the given chain id (number or CAIP-2 string) matches the payment chain.
- */
+/** True when `chainId` is the configured payment chain (numeric id, numeric string, or CAIP-2). */
 export function isChainIdBase(chainId: string | number): boolean {
   if (typeof chainId === "number") return chainId === BASE_CHAIN_ID;
   return chainId === BASE_CHAIN_ID_CAIP2 || chainId === String(BASE_CHAIN_ID);
@@ -23,7 +17,6 @@ export const privyConfig: PrivyClientConfig = {
   appearance: {
     theme: "dark",
     accentColor: "#22c55e",
-    logo: undefined,
     walletList: ["metamask", "coinbase_wallet", "base_account", "rainbow"],
   },
   embeddedWallets: {
